requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.4
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
appnope==0.1.4 ; sys_platform == 'darwin'
    # via ipykernel
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via aiohttp
blinker==1.9.0
    # via flask
cachelib==0.9.0
    # via flask-caching
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1 ; implementation_name == 'pypy'
    # via pyzmq
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   ipython
    #   loguru
    #   tqdm
comm==0.2.2
    # via ipykernel
dash==2.18.2
    # via
    #   dash-bootstrap-components
    #   michelin-guide-restaurants-dashboard
dash-bootstrap-components==1.6.0
    # via michelin-guide-restaurants-dashboard
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
dataclasses-json==0.6.7
    # via langchain-community
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
distlib==0.3.9
    # via virtualenv
distro==1.9.0
    # via openai
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via virtualenv
flask==3.0.3
    # via
    #   dash
    #   flask-caching
flask-caching==2.3.0
    # via michelin-guide-restaurants-dashboard
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.2 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
gunicorn==23.0.0
    # via michelin-guide-restaurants-dashboard
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via dash
ipykernel==6.29.5
ipython==9.2.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
itsdangerous==2.2.0
    # via flask
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via flask
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
langchain==0.3.4
    # via
    #   langchain-community
    #   michelin-guide-restaurants-dashboard
langchain-community==0.3.3
    # via michelin-guide-restaurants-dashboard
langchain-core==0.3.63
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.2.3
    # via michelin-guide-restaurants-dashboard
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
loguru==0.7.2
    # via michelin-guide-restaurants-dashboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.26.1
    # via dataclasses-json
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==1.41.0
    # via plotly
nest-asyncio==1.6.0
    # via
    #   dash
    #   ipykernel
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   pandas
openai==1.82.1
    # via langchain-openai
orjson==3.10.18 ; platform_python_implementation != 'PyPy'
    # via langsmith
packaging==24.2
    # via
    #   gunicorn
    #   ipykernel
    #   langchain-core
    #   marshmallow
    #   plotly
pandas==2.2.3
    # via michelin-guide-restaurants-dashboard
parso==0.8.4
    # via jedi
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via ipython
platformdirs==4.3.8
    # via
    #   jupyter-core
    #   virtualenv
plotly==6.1.2
    # via dash
pre-commit==4.2.0
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycparser==2.22 ; implementation_name == 'pypy'
    # via cffi
pydantic==2.9.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   michelin-guide-restaurants-dashboard
    #   openai
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pandas
python-dotenv==1.0.1
    # via
    #   michelin-guide-restaurants-dashboard
    #   pydantic-settings
pytz==2025.2
    # via pandas
pywin32==310 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
    # via jupyter-core
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   pre-commit
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   dash
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
retrying==1.3.4
    # via dash
ruff==0.11.12
setuptools==80.9.0
    # via dash
six==1.17.0
    # via
    #   python-dateutil
    #   retrying
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
stack-data==0.6.3
    # via ipython
tenacity==9.1.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via langchain-openai
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.13.2
    # via
    #   dash
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
virtualenv==20.31.2
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.0.6
    # via
    #   dash
    #   flask
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
yarl==1.20.0
    # via aiohttp
zipp==3.22.0
    # via importlib-metadata
seaborn
dash
dash-bootstrap-components
plotly
pandas
